Make assignment as void
Stop automatic duplication for assignment value,
because of not suit for wasm instruction.

WASM instructions doesn't have `duplicate`
a value on stack,
so it is needed to allocate local variable for temporary value.
But assigned value is not always need, so it is wasteful.

Handle assignment operator as `void`.
If multiple assignment expression (ex. `a = b = 0`) is used,
then replace it using comma expression:

```c
(tmp = r, l = tmp, tmp)
```