Avoid source generation conflict
libsrc/_wasm/crt0.c and libc.c in are built from make target `update-wcc-lib.
It is triggered when any source file is updated.

When make runs in parallel (`-j4`),
it is triggered 2 times in parallel, and may be conflict.

Separate source generation to avoid it.

Also update lib_list.json from those files and described it as make-target.